# src/scripts/api-spec.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Customers Lineup Api",
    "version": "v1"
  },
  "paths": {
...
    "/api/query/user": {
      "get": {
        "tags": [
          "Query"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CLup.API.Contracts.Users.GetUser.GetUserResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CLup.API.Contracts.Users.GetUser.GetUserResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CLup.API.Contracts.Users.GetUser.GetUserResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/query/user/business/{businessId}": {
      "get": {
        "tags": [
          "Query"
        ],
        "parameters": [
          {
            "name": "BusinessId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CLup.API.Contracts.Businesses.GetBusiness.GetBusinessResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CLup.API.Contracts.Businesses.GetBusiness.GetBusinessResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CLup.API.Contracts.Businesses.GetBusiness.GetBusinessResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/query/user/businesses": {
      "get": {
        "tags": [
          "Query"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CLup.API.Contracts.Users.GetUser.GetUserResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CLup.API.Contracts.Users.GetUser.GetUserResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CLup.API.Contracts.Users.GetUser.GetUserResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/query/business/all": {
      "get": {
        "tags": [
          "Query"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CLup.API.Contracts.Businesses.GetAllBusinessesResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CLup.API.Contracts.Businesses.GetAllBusinessesResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CLup.API.Contracts.Businesses.GetAllBusinessesResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/query/user/notemployedbybusiness/{businessId}": {
      "get": {
        "tags": [
          "Query"
        ],
        "parameters": [
          {
            "name": "BusinessId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CLup.API.Contracts.Users.UsersNotEmployedByBusiness.UsersNotEmployedByBusinessResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CLup.API.Contracts.Users.UsersNotEmployedByBusiness.UsersNotEmployedByBusinessResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CLup.API.Contracts.Users.UsersNotEmployedByBusiness.UsersNotEmployedByBusinessResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/timeslot": {
      "post": {
        "tags": [
          "TimeSlot"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CLup.API.Contracts.TimeSlots.GenerateTimeSlots.GenerateTimeSlotsRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CLup.API.Contracts.TimeSlots.GenerateTimeSlots.GenerateTimeSlotsRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CLup.API.Contracts.TimeSlots.GenerateTimeSlots.GenerateTimeSlotsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/timeslot/{timeSlotId}": {
      "delete": {
        "tags": [
          "TimeSlot"
        ],
        "parameters": [
          {
            "name": "timeSlotId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "businessId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CLup.API.Contracts.Businesses.GetAllBusinessesResponse": {
        "type": "object",
        "properties": {
          "businesses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CLup.Application.Businesses.BusinessDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CLup.API.Contracts.Businesses.GetBusiness.GetBusinessResponse": {
        "type": "object",
        "properties": {
          "business": {
            "$ref": "#/components/schemas/CLup.Application.Businesses.BusinessDto"
          }
        },
        "additionalProperties": false
      },
...
      "CLup.API.Contracts.TimeSlots.GenerateTimeSlots.GenerateTimeSlotsRequest": {
        "type": "object",
        "properties": {
          "businessId": {
            "type": "string",
            "format": "uuid"
          },
          "date": {
            "$ref": "#/components/schemas/System.DateOnly"
          }
        },
        "additionalProperties": false
      },
      "CLup.API.Contracts.Users.GetUser.GetUserResponse": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/CLup.Application.Users.UserDto"
          }
        },
        "additionalProperties": false
      },
...
      "CLup.API.Contracts.Users.UsersNotEmployedByBusiness.UsersNotEmployedByBusinessResponse": {
        "type": "object",
        "properties": {
          "businessId": {
            "type": "string",
            "format": "uuid"
          },
          "users": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CLup.Application.Users.UserDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CLup.Application.Bookings.BookingDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "timeSlotId": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "businessId": {
            "type": "string",
            "format": "uuid"
          },
          "business": {
            "type": "string",
            "nullable": true
          },
          "longitude": {
            "type": "number",
            "format": "double"
          },
          "latitude": {
            "type": "number",
            "format": "double"
          },
          "street": {
            "type": "string",
            "nullable": true
          },
          "userEmail": {
            "type": "string",
            "nullable": true
          },
          "interval": {
            "type": "string",
            "nullable": true
          },
          "date": {
            "type": "string",
            "nullable": true
          },
          "capacity": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CLup.Application.Businesses.BusinessDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "ownerId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "zip": {
            "type": "integer",
            "format": "int32"
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "street": {
            "type": "string",
            "nullable": true
          },
          "longitude": {
            "type": "number",
            "format": "double"
          },
          "latitude": {
            "type": "number",
            "format": "double"
          },
          "opens": {
            "type": "string",
            "nullable": true
          },
          "closes": {
            "type": "string",
            "nullable": true
          },
          "timeSlotLength": {
            "type": "integer",
            "format": "int32"
          },
          "businessHours": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "$ref": "#/components/schemas/CLup.Domain.Businesses.Enums.BusinessType"
          },
          "capacity": {
            "type": "integer",
            "format": "int32"
          },
          "bookings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CLup.Application.Bookings.BookingDto"
            },
            "nullable": true
          },
          "employees": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CLup.Application.Employees.EmployeeDto"
            },
            "nullable": true
          },
          "receivedMessages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CLup.Application.Messages.MessageDto"
            },
            "nullable": true
          },
          "sentMessages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CLup.Application.Messages.MessageDto"
            },
            "nullable": true
          },
          "timeSlots": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CLup.Application.TimeSlots.TimeSlotDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CLup.Application.Employees.EmployeeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "businessId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "privateEmail": {
            "type": "string",
            "nullable": true
          },
          "companyEmail": {
            "type": "string",
            "nullable": true
          },
          "business": {
            "type": "string",
            "nullable": true
          },
          "employedSince": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CLup.Application.Messages.MessageDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "senderId": {
            "type": "string",
            "format": "uuid"
          },
          "receiverId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "content": {
            "type": "string",
            "nullable": true
          },
          "date": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CLup.Application.TimeSlots.TimeSlotDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "businessId": {
            "type": "string",
            "format": "uuid"
          },
          "business": {
            "type": "string",
            "nullable": true
          },
          "date": {
            "type": "string",
            "nullable": true
          },
          "start": {
            "type": "string",
            "nullable": true
          },
          "end": {
            "type": "string",
            "nullable": true
          },
          "interval": {
            "type": "string",
            "nullable": true
          },
          "capacity": {
            "type": "string",
            "nullable": true
          },
          "available": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "CLup.Application.Users.UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "zip": {
            "type": "integer",
            "format": "int32"
          },
          "street": {
            "type": "string",
            "nullable": true
          },
          "longitude": {
            "type": "number",
            "format": "double"
          },
          "latitude": {
            "type": "number",
            "format": "double"
          },
          "role": {
            "type": "string",
            "nullable": true
          },
          "bookings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CLup.Application.Bookings.BookingDto"
            },
            "nullable": true
          },
          "businessIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "nullable": true
          },
          "receivedMessages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CLup.Application.Messages.MessageDto"
            },
            "nullable": true
          },
          "sentMessages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CLup.Application.Messages.MessageDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CLup.Domain.Businesses.Enums.BusinessType": {
        "enum": [
          1,
          2,
          3,
          4,
          5,
          6,
          7
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "Microsoft.AspNetCore.Mvc.ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "System.DateOnly": {
        "type": "object",
        "properties": {
          "year": {
            "type": "integer",
            "format": "int32"
          },
          "month": {
            "type": "integer",
            "format": "int32"
          },
          "day": {
            "type": "integer",
            "format": "int32"
          },
          "dayOfWeek": {
            "$ref": "#/components/schemas/System.DayOfWeek"
          },
          "dayOfYear": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "dayNumber": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "System.DayOfWeek": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5,
          6
        ],
        "type": "integer",
        "format": "int32"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\"",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}